policy_gradient/requirements.txt
# Neural Network Model Dependencies
pytorch

# Core Python Libraries
numpy

# Visualization Tools
tensorboard
matplotlib
imageio

# Environments
gymnasium